fix a float bug
- fix a float bug
add (float), otherwise truncated.
- remove unused functions.

+++ ZoningPlanner/MCMCUtil.cpp
@@ -25,7 +25,7 @@ void MCMCUtil::computeFeature(int city_size, int num_features, vector<uchar>& zo
 		// ガウス分布を使ってみよう
 		float K = 1.6f;
 		float sigma = (float)city_size * 0.25;
-		feature[i] = K * expf(-dist[i][s]*dist[i][s]/2/sigma/sigma);
+		feature[i] = K * expf(-(float)dist[i][s]*(float)dist[i][s]/2.0/sigma/sigma);
 	}
 }
